+++ src/score/engine/span-props.ts
@@ -1,21 +1,13 @@
 import { AnnotationKind } from "../pub";
-import { MusicObject } from "./music-object";
 import { ObjRhythmColumn } from "./obj-rhythm-column";
-import { ObjMeasure } from "./obj-measure";
 import { LayoutObjectWrapper } from "./layout-object";
 import { SpanStopObject, ObjSpanSegment } from "./obj-span-segment";
 import { ObjAnnotation } from "./obj-annotation";
 import { Assert } from "@tspro/ts-utils-lib";
 
-export type ExtensionLineStyle = "solid" | "dashed";
-export type ExtensionLinePos = "bottom" | "middle";
-
-export function getExtensionLineAnchorY(linePos: ExtensionLinePos) {
-    switch (linePos) {
-        case "bottom": return 0.8;
-        case "middle": return 0.5;
-    }
-}
+export enum SpanType { Line, Hairpin }
+export type SpanLineStyle = "solid" | "dashed";
+export type SpanLineVPos = "bottom" | "middle";
 
 export class SpanRange {
     public readonly columnRange: ObjRhythmColumn[];
@@ -45,12 +37,14 @@ export class SpanProps {
     private readonly length: number;
     private readonly visible: boolean;
 
-    private readonly lineStyle: ExtensionLineStyle;
-    private readonly linePos: ExtensionLinePos;
+    readonly type: SpanType;
+
+    readonly lineStyle: SpanLineStyle;
+    readonly linePos: SpanLineVPos;
 
     private readonly startColumn: ObjRhythmColumn;
 
-    constructor(readonly headObj: LayoutObjectWrapper, startColumn: ObjRhythmColumn, length: number, visible: boolean, lineStyle: ExtensionLineStyle, linePos: ExtensionLinePos) {
+    constructor(readonly headObj: LayoutObjectWrapper, startColumn: ObjRhythmColumn, length: number, visible: boolean) {
         Assert.assert(headObj.musicObj instanceof ObjAnnotation, "Head object must be annotation!");
 
         this.annotationObj = headObj.musicObj as ObjAnnotation;
@@ -59,8 +53,12 @@ export class SpanProps {
         this.length = length + 1; // + 1 just to connect with following elem.
         this.visible = visible;
 
-        this.lineStyle = lineStyle;
-        this.linePos = linePos;
+        this.type = this.annotationObj.kind === "<" || this.annotationObj.kind === ">"
+            ? SpanType.Hairpin
+            : SpanType.Line;
+
+        this.lineStyle = "dashed";
+        this.linePos = "bottom";
 
         this.startColumn = startColumn;
     }
@@ -71,14 +69,6 @@ export class SpanProps {
 
     isVisible() {
         return this.visible;
-    }
-
-    getLineStyle(): ExtensionLineStyle {
-        return this.lineStyle;
-    }
-
-    getLinePos(): ExtensionLinePos {
-        return this.linePos;
     }
 
     private static StopNavigations = [AnnotationKind.EndRepeat, AnnotationKind.Ending];